Layout and style reworks, tree view, switchable players & more
- Changed user interface colors to be in line with all my other
neko-styled configurations and application themes.
- Incorporated some of @TheQwertiest CaTRoX fork changes and code
including switchable player sizes for mini and micro players as well as
always on top functionality.
- Made customizing and resizing panels more straight forward with better
padding and margins.
- Panels should now have consistent colors and spacing all over.
- Updated the biography view to the newest version which enables viewing
of artist images again.
- Added a tree style panel alongside the default library filters.
- Cleaned up the repository and split files into the NekoRoX theme
directory for better management.
- Added another quality of life plugin via the Discord Rich Presence
plugin allowing live broadcasting of users' current playing song.
- Fixed the YouTube search field sometimes getting lost when resizing
panels after moving it to the bottom of the playlist splitter.
- Indirectly fixed #2 by moving all menu items into a single menu. The
reason for this is to keep the menu consistent across all player sizes
and modes.